feat: searchbar
Updated folder structure for searchbar-ui component. Refactored code to make it more readable
