+++ src/app/features/ui/index.ts
@@ -1,2 +1,3 @@
-export * from "./searchbar/searchbar.component";
-export * from "./searchbar/searchbar-dropdown.component";
+export * from "./searchbar-ui/searchbar/searchbar.component";
+export * from "./searchbar-ui/search.types";
+export * from "./searchbar-ui/searchbar-dropdown/searchbar-dropdown.component";
