+++ src/app/features/ui/searchbar-ui/search.types.ts
@@ -1,0 +1,11 @@
+export type SearchType = {
+  searchTypeName: string;
+  searchTypeIcon: string;
+  isActive: boolean;
+};
+
+export type SearchResult = {
+  searchResultTitle: string;
+  searchResultPhrase?: string;
+  searchResultImg?: string;
+};
