+++ src/app/features/ui/searchbar-ui/searchbar-dropdown/searchbar-dropdown.component.scss
