+++ src/app/features/ui/searchbar-ui/searchbar/searchbar.component.scss
